+++ churchinfo/PaddleNumList.php
@@ -21,7 +21,7 @@ $linkBack = FilterInput($_GET["linkBack"]);
 
 $iFundRaiserID = $_SESSION['iCurrentFundraiser'];
 
-if ($iFundRaiserID) {
+if ($iFundRaiserID > 0) {
 	//Get the paddlenum records for this fundraiser
 	$sSQL = "SELECT pn_ID, pn_fr_ID, pn_Num, pn_per_ID,
 	                a.per_FirstName as buyerFirstName, a.per_LastName as buyerLastName
@@ -29,6 +29,8 @@ if ($iFundRaiserID) {
 	         LEFT JOIN person_per a ON pn_per_ID=a.per_ID
 	         WHERE pn_FR_ID = '" . $iFundRaiserID . "' ORDER BY pn_Num"; 
 	 $rsPaddleNums = RunQuery($sSQL);
+} else {
+	$rsPaddleNums = 0;
 }
 
 $sPageTitle = gettext("Buyers for this fundraiser:");
@@ -37,7 +39,7 @@ require "Include/Header.php";
 
 <?php
 echo "<form method=\"post\" action=\"Reports/FundRaiserStatement.php?CurrentFundraiser=$iFundRaiserID&linkBack=FundRaiserEditor.php?FundRaiserID=$iFundRaiserID&CurrentFundraiser=$iFundRaiserID\">\n";
-if ($iFundRaiserID)
+if ($iFundRaiserID > 0)
 	echo "<input type=button class=icButton value=\"".gettext("Select all")."\" name=SelectAll onclick=\"javascript:document.location='PaddleNumList.php?CurrentFundraiser=$iFundRaiserID&SelectAll=1&linkBack=FundRaiserEditor.php?FundRaiserID=$iFundRaiserID&CurrentFundraiser=$iFundRaiserID';\">\n";
 	echo "<input type=button class=icButton value=\"".gettext("Select none")."\" name=SelectNone onclick=\"javascript:document.location='PaddleNumList.php?CurrentFundraiser=$iFundRaiserID&linkBack=FundRaiserEditor.php?FundRaiserID=$iFundRaiserID&CurrentFundraiser=$iFundRaiserID';\">\n";
 	echo "<input type=button class=icButton value=\"".gettext("Add Buyer")."\" name=AddBuyer onclick=\"javascript:document.location='PaddleNumList.php?CurrentFundraiser=$iFundRaiserID&linkBack=FundRaiserEditor.php?FundRaiserID=$iFundRaiserID&CurrentFundraiser=$iFundRaiserID';\">\n";
@@ -58,31 +60,33 @@ if ($iFundRaiserID)
 $tog = 0;
 
 //Loop through all buyers
-while ($aRow =mysql_fetch_array($rsPaddleNums))
-{
-	extract($aRow);
-
-	$sRowClass = "RowColorA";
-?>
-	<tr class="<?php echo $sRowClass ?>">
-		<td>
-			<input type="checkbox" name="Chk<?php echo $pn_ID."\""; if (isset($_GET["SelectAll"])) echo " checked=\"yes\"";?>></input>
-		</td>
-		<td>
-			<?php echo $pn_Num?>&nbsp;
-		</td>
-		<td>
-			<?php echo $buyerFirstName . " " . $buyerLastName ?>&nbsp;
-		</td>
-		<td>
-			<a href="PaddleNumEditor.php?PaddleNumID=<?php echo $pn_ID . "&linkBack=PaddleNumList.php";?>">Edit</a>
-		</td>
-		<td>
-			<a href="PaddleNumDelete.php?PaddleNumID=<?php echo $pn_ID . "&linkBack=PaddleNumList.php?FundRaiserID=" . $iFundRaiserID;?>">Delete</a>
-		</td>
-	</tr>
-<?php
-} // while
+if ($rsPaddleNums) {
+	while ($aRow =mysql_fetch_array($rsPaddleNums))
+	{
+		extract($aRow);
+	
+		$sRowClass = "RowColorA";
+	?>
+		<tr class="<?php echo $sRowClass ?>">
+			<td>
+				<input type="checkbox" name="Chk<?php echo $pn_ID."\""; if (isset($_GET["SelectAll"])) echo " checked=\"yes\"";?>></input>
+			</td>
+			<td>
+				<?php echo $pn_Num?>&nbsp;
+			</td>
+			<td>
+				<?php echo $buyerFirstName . " " . $buyerLastName ?>&nbsp;
+			</td>
+			<td>
+				<a href="PaddleNumEditor.php?PaddleNumID=<?php echo $pn_ID . "&linkBack=PaddleNumList.php";?>">Edit</a>
+			</td>
+			<td>
+				<a href="PaddleNumDelete.php?PaddleNumID=<?php echo $pn_ID . "&linkBack=PaddleNumList.php?FundRaiserID=" . $iFundRaiserID;?>">Delete</a>
+			</td>
+		</tr>
+	<?php
+	} // while
+} // if
 ?>
 
 </table>
